requirements.txt
lxml
requests
streamlit
google-generativeai
python-dotenv
PyQt6
PyQt6-WebEngine